+++ src/reducer.js
@@ -2,7 +2,8 @@
 export default function (state = {}, action) {
   switch(action.type) {
     case 'f1-sticky/SCROLL':
-      return {...state, [action.meta.name]: action.payload}
+    case 'f1-sticky/RESIZE':
+      return {...state, [action.meta.name]: {...state[action.meta.name], ...action.payload}}
   }
   return state
 }