+++ test/child.spec.js
@@ -30,6 +30,21 @@ describe('StickyChildComponent', () => {
     })
   })
 
+  // this does not actually work, because jsdom will not calculate the actual height of the
+  // node.  We'd need to use something like phantomjs to run the tests.
+  it('does not sticky when taller than parent', (done) => {
+    const wrapper = mount(<Component parentHeight={1000} parentScrollTop={0}><div style={{height: '1000px', width: '100px'}} className='mydiv' /></Component>)
+    wrapper.setProps({
+      parentScrollTop: 123,
+      parentHeight: 200
+    }, () => {
+      const node = wrapper.find('.mydiv').get(0).parentNode
+      expect(node.style.position).to.equal('')
+      expect(node.style.top).to.equal('')
+      done()
+    })
+  })
+
   it('sets spacer margin when scrolled past the top', (done) => {
     const wrapper = mount(<Component parentScrollTop={0}><div className='mydiv' /></Component>)
     wrapper.setProps({