+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog f1-sticky
 
+## 1.0.2
+
+ * Prevent sticking when taller than parent
+
+## 1.0.1
+
+ * Remove sticky from state
+
 ## 1.0.0
 
-* Initial version
+ * Initial version